Make Database a class and use it in the application to get submissions
Note that in app/main.rb I'm just assigning Database.new to a variable
(`db = Database.new`), rather than a class attribute (`@db = Database.new`).
This is because of the way that Sinatra's scopes work - inside the `get '/'`
block the code doesn't have access to the attributes on the Sinatra::Application
instance.