Set up the application to be easier to test
We want the "main" file to be able to talk to the "database" file, but
in our tests we'd like to "mock" the behaviour of the database so the
tests can run independently (that is: without having a database).

To achieve this, we've set up the application to create an instance of
the `Database` class, and pass this in to the constructor of the
`ApprenticeNews` class. This means in the tests we can set up the
application so that it passes in a "double" (which is kind of like a
mock) instead of the real database. This means we can write the tests
for the website independently of the Database class.